Move timeout-sensitive tests out of the unit suite
Three tests asserted on wall-clock behavior and so failed under CPU load
rather than on any defect:

  - GraphColoringScalingTester (every algorithm, under a hard timeout)
  - test_deterministic_algorithm_colors_dense_graph (asserts termination
    inside PER_ALGO_TIMEOUT)
  - TestTimedBlock.test_timer (asserts a sleep(0.5) is measured within 0.2s)

On a busy machine a healthy algorithm gets killed on its budget and the test
reports a failure that says nothing about the code; observed failing 6 of
these on unmodified develop with the box loaded. They now live under
test/performance/, which is not part of the ordinary unit run.

The graph fixtures and assess_coloring are shared by the correctness suite
and the scaling suite, so they move to test/helpers/graphcoloring_graphs.py
rather than being duplicated -- assess_coloring decides whether a coloring is
correct, and a fix to it must not land in only one of the two.

CI keeps the same coverage: the change detector now also watches the
relocated suite and the shared fixtures, and the conditional slow-test step
runs the new path. The stale pygsti/tools/graphcoloring.py entry is dropped
from the detector, since #848 turned that module into a subpackage.
