+++ .github/workflows/reuseable-main.yml
@@ -133,9 +133,11 @@ jobs:
         COV_FLAG: ${{ inputs.run-coverage == 'true' && '--cov=pygsti' || '' }}
       run: |
         python -Ic "import pygsti; print(pygsti.__version__); print(pygsti.__path__)"
-        # Slow/benchmark-style tests (currently just the graphcoloring dense-graph
-        # suite) are excluded here and run in their own step below instead, so a
-        # slow/flaky run there can't eat into or fail the main unit-test budget.
+        # Wall-clock-sensitive suites live under test/performance/ and so are not
+        # collected here at all; the graphcoloring scaling suite is run
+        # conditionally in its own step below, where a slow run can't eat into or
+        # fail the main unit-test budget. The `-m "not slow"` filter is kept as a
+        # backstop in case a benchmark-style test lands under test/unit again.
         python -m pytest -n auto --dist loadscope -m "not slow" $COV_FLAG --durations=25 test/unit
     - name: Detect graphcoloring net changes
       # The graphcoloring scaling tests are slow, so only run them when this branch
@@ -173,9 +175,10 @@ jobs:
         git fetch --no-tags --depth=1 origin "${DIFF_BASE}"
 
         if git diff --quiet "origin/${DIFF_BASE}" HEAD -- \
-          pygsti/tools/graphcoloring.py \
           pygsti/tools/graphcoloring \
-          test/unit/tools/test_graphcoloring.py
+          test/unit/tools/test_graphcoloring.py \
+          test/performance/test_graphcoloring_scaling.py \
+          test/helpers/graphcoloring_graphs.py
         then
           echo "changed=false" >> "$GITHUB_OUTPUT"
         else
@@ -189,7 +192,7 @@ jobs:
         GRAPHCOLORING_CHANGED: ${{ steps.graphcoloring.outputs.changed }}
       run: |
         python -Ic "import pygsti; print(pygsti.__version__); print(pygsti.__path__)"
-        python -m pytest -m slow --durations=25 test/unit/tools/test_graphcoloring.py
+        python -m pytest -m slow --durations=25 test/performance/test_graphcoloring_scaling.py
     - name: Run workflow tests
       if: ${{ inputs.run-integration-tests == 'true' }}
       shell: bash
